Add Screen Reader/Clicker
Adds a python script to automatically read cards from the screen, input
them to the solver, and click on the correct cards in order.

Also, fix a small-but-important bug the previous formatting commit
introduced to the C++ program that caused the transposition table to not
store state between steps. This severely hampered performance.